+++ src/main/java/de/alpharogroup/gson/strategy/AbstractExclusionStrategy.java
@@ -1,0 +1,22 @@
+package de.alpharogroup.gson.strategy;
+
+import com.google.gson.ExclusionStrategy;
+import com.google.gson.FieldAttributes;
+
+import java.lang.annotation.Annotation;
+
+public abstract class AbstractExclusionStrategy implements ExclusionStrategy
+{
+
+	@Override public boolean shouldSkipClass(Class<?> clazz)
+	{
+		return false;
+	}
+
+	@Override public boolean shouldSkipField(FieldAttributes field)
+	{
+		return field.getAnnotation(getAnnotationClass()) != null;
+	}
+	
+	public abstract Class<? extends Annotation> getAnnotationClass();
+}