+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 1.2-SNAPSHOT
 -------------
 
+ADDED: 
+
+- new abstract class for the interface ExclusionStrategy created
+
 Version 1.1
 -------------